clean: separate Spokesman component
